Fix userEntity layer and create tests for all the servises

+++ Backend/src/main/java/com/codecool/backend/controller/dto/UserEntityProfileDto.java
@@ -2,5 +2,5 @@ package com.codecool.backend.controller.dto;
 
 import java.math.BigDecimal;
 
-public record UserEntityProfileDto(int id, String username, String email, BigDecimal targetAmount) {
+public record UserEntityProfileDto(Long id, String username, String email, BigDecimal targetAmount) {
 }
